Title: 세 수의 합과 평균 구하기, Time: 105ms, Memory: 9MB, Status: Passed - CodeTree

+++ 240725/세 수의 합과 평균 구하기/sum-and-mean-of-three-numbers.java
@@ -9,6 +9,6 @@ public class Main {
         int b = Integer.parseInt(st.nextToken());
         int c = Integer.parseInt(st.nextToken());
 
-        System.out.printf("%d\n%.0f", (a + b + c), (double)(a + b + c) / 3);
+        System.out.printf("%d\n%d", (a + b + c), (a + b + c) / 3);
     }
 }
